ui : move mention-badge constants into $lib/constants
